+++ tests/_data/snapshots/get_bom_with_tools_with_component_and_service_migrate-1.2.xml.bin
